+++ inc/core/Parser/Activity/Common/Data/Metadata.php
@@ -81,7 +81,7 @@ class Metadata
      */
     public function interpretTimestampAsServerTime($timestamp)
     {
-        $this->TimezoneOffset = round((new \DateTime())->setTimestamp($timestamp)->getOffset() / 60);
+        $this->TimezoneOffset = (int)round((new \DateTime())->setTimestamp($timestamp)->getOffset() / 60);
         $this->Timestamp = LocalTime::fromServerTime($timestamp)->getTimestamp();
     }
 
@@ -95,7 +95,7 @@ class Metadata
 
             $this->setTimestamp(
                 $dateTime->getTimestamp() + $dateTime->getOffset(),
-                round($dateTime->getOffset() / 60)
+                (int)round($dateTime->getOffset() / 60)
             );
         } catch (\Exception $e) {
             // Invalid date
@@ -112,7 +112,7 @@ class Metadata
 
             $this->setTimestamp(
                 $localTimestamp,
-                round((new \DateTime())->setTimestamp($localTimestamp)->getOffset() / 60)
+                (int)round((new \DateTime())->setTimestamp($localTimestamp)->getOffset() / 60)
             );
         } else {
             $this->setTimestampAndTimezoneOffsetFrom($string);
