+++ inc/draw/class.PlotWeekSumData.php
@@ -28,7 +28,7 @@ class PlotWeekSumData extends PlotSumData {
 			$this->timerEnd = 26;
 		} else {
 			if (Request::param('y') == self::LAST_12_MONTHS) {
-				$yearEnd = date('Y') - 1;
+				$yearEnd = (int)date('Y') - 1;
 			} else {
 				$yearEnd = (int)Request::param('y');
 			}
@@ -62,7 +62,7 @@ class PlotWeekSumData extends PlotSumData {
 	protected function getXLabels() {
 		$weeks = array();
 		$add = ($this->Year == parent::LAST_6_MONTHS || $this->Year == parent::LAST_12_MONTHS) ? 0 : $this->WeekStart->phpWeek() - $this->timerEnd;
-		$yearDiff = $add == 0 ? 0 : date('Y') - $this->Year;
+		$yearDiff = $add == 0 ? 0 : (int)date('Y') - $this->Year;
 
 		for ($w = $this->timerStart; $w <= $this->timerEnd; $w++) {
 			$time = strtotime($this->WeekStart->lastDayOfWeekForStrtotime()." -".($this->timerEnd - $w + $add)." weeks -".$yearDiff." years");
